Gravity added

+++ SolarSystemProject/SolarSystemProject/CelestialEntity.cpp
@@ -1,7 +1,9 @@
 #include "CelestialEntity.h"
 #include <iostream>
+#include <SFML/Graphics.hpp>
+#include "SolarSpace.h"
 #include <string>
-CelestialEntity::CelestialEntity(int size, int position[2], float initialVelocity)
+CelestialEntity::CelestialEntity(float size, float position[2], float initialVelocity)
 {
 	this->size = size;
 	this->position[0] = position[0];
@@ -13,10 +15,11 @@ void CelestialEntity::setType(std::string type)
 {
 	this->type = type;
 }
-int CelestialEntity::getSize() const
+float CelestialEntity::getSize() const
 {
 	return size;
 }
+
 
 void CelestialEntity::CheckVelocitybyType()
 {
